rename and repackage part 2
