+++ services/src/main/java/io/precognito/services/storage/IndexingFactoryConverter.java
@@ -1,7 +1,6 @@
-package io.precognito.services;
+package io.precognito.services.storage;
 
 import io.precognito.services.fixture.FixturedIndexerService;
-import io.precognito.services.storage.StorageIndexer;
 import org.eclipse.microprofile.config.spi.Converter;
 
 public class IndexingFactoryConverter implements Converter<StorageIndexer> {
