--- services/src/main/resources/application.properties
@@ -1,40 +1,0 @@
-# DEV Configuration file
-# key = value
-
-# used to configured how real the test is;
-# TEST = no connectivity; DEV = connected; PROD = connected
-#
-cloud.region=eu-west-2
-storage.uploader=TEST
-storage.indexer=TEST
-storage.query=TEST
-
-%dev.storage.uploader=AWS
-%dev.storage.indexer=AWS
-
-%test.storage.uploader=TEST
-%test.storage.indexer=TEST
-
-
-
-quarkus.dynamodb.endpoint-override=http://192.168.99.100:8000
-#quarkus.dynamodb.aws.region=eu-central-1
-#quarkus.dynamodb.aws.credentials.type=static
-#quarkus.dynamodb.aws.credentials.static-provider.access-key-id=test-key
-#quarkus.dynamodb.aws.credentials.static-provider.secret-access-key=test-secret
-
-
-
-# DEBUG console logging
-quarkus.log.console.enable=true
-quarkus.log.console.format=%d{HH:mm:ss} %-5p [%c] %s%e%n
-quarkus.log.console.level=TRACE
-
-# TRACE file logging
-quarkus.log.file.enable=true
-quarkus.log.file.path=/tmp/quarkus.log
-quarkus.log.file.level=TRACE
-quarkus.log.file.format=%d{HH:mm:ss} %-5p [%c{2.}]] (%t) %s%e%n
-
-# custom loggers
-quarkus.log.category."com.logscapeng".level=INFO
