+++ services/src/test/resources/META-INF/services/org.eclipse.microprofile.config.spi.Converter
@@ -1,4 +1,4 @@
 io.precognito.services.storage.StorageFactoryConverter
-io.precognito.services.IndexingFactoryConverter
+io.precognito.services.storage.IndexingFactoryConverter
 io.precognito.services.query.QueryFactoryConverter
 io.precognito.services.search.SearchFactoryConverter
